Now should work also on msys platform
For some reason, on msys ($^O == msys) when a line is processed,
chomp() doesn't remove a trailing "\r" character. This is in
contrast to what happens on MSWin32.

I decided to put a kludge fix in, so when we are on msys,
we explicitly strip the trailing newline characters from every line.

Only on msys, we also cannot check for perfect roundtrip output
because a newline gets appended to the end of the files (??)